iot-simulation/requirements.txt
requests>=2.25.1
pyjwt>=2.1.0
cryptography>=3.4.7